requirements.txt
aioice==0.9.0
aiortc==1.11.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
captcha==0.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
extra-streamlit-components==0.1.71
fastapi==0.115.11
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
Janome==0.5.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
narwhals==1.30.0
netifaces==0.11.0
numpy==2.2.4
ollama==0.4.8
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
PyJWT==2.10.1
pylibsrtp==0.12.0
pyOpenSSL==25.0.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
streamlit==1.43.2
streamlit-audiorecorder==0.0.6
streamlit-authenticator==0.4.2
streamlit-webrtc==0.60.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
